backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
httpx
beautifulsoup4
openai
python-dotenv
